split app logic in different directories. add route
